django/requirements.txt
Django>=4.2
gunicorn
grpcio
grpcio-tools
requests
django-cors-headers
djangorestframework